refactor(services): split cell/abilities.rs into submodules
Carve the 931-line abilities.rs into a directory laid out by lifecycle
stage: ground-target dispatch, the main use_ability flow, entity-method
routing, on-death loot drop, and the deterministic combat RNG.

Public API unchanged. No behavior changes.
